Boot a bunch of code out of Value.hs.
Create a SAWScript.ValueOps module with a bunch of code from
SAWCentral.Value that doesn't need to be in SAWCentral; this is
stuff that was only actually used from the saw-script library.

Note that pretty much all of Value.hs ought not to be in SAWCentral,
because it's the SAWScript notion of values (which probably should not
be shared with the remote API) combined with the SAWScript
interpreter's internal state and monads (which definitely should not
be) but there is a long way to go before either of those things can be
shifted out.

Provided that we don't actually need a common notion of values shared
between the scripting interfaces, it's ok that the Value type and
interpreter state depend on each other. However, there are references
to the Value type in assorted places there shouldn't be by that
criterion, and of course the interpreter state is splattered
everywhere.

Even before that there's a bunch more code that could go to ValueOps
except for specific entanglements that can and should get cleaned up.

Also note that some of this code should probably in fact be in
Interpreter.hs or other places. For the time being I'm just dumping it
all; we can reorganize the saw-script library once we get it
adequately partitioned from the rest of the system. It's probably
premature to try before that point.
